Add files via upload
The cards now slide smoothly when opening. A row of cards seems to snap when a card in the row above it opens, but the row will slide smoothly back into place when the aforementioned card closes.  Let's see if we can get it to slide smoothly when the card opens as well. 

Another issue here is that even though the ingredients list isn't visible when the card is closed (its opacity is 0), it's still technically displaing (its display is set to block. Removing the display:none property is what enabled the smooth sliding. I'm going to try and use javascript to add display:none to the ingredients list *after* it's done closing and setting its display to block right before it opens, so that it still slide up and down smoothly.